# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KLA Katingan API Documentation",
        "description": "API Documentation for KLA Katingan",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Environment"
        }
    ],
    "paths": {
...
        "/api/news": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Membuat berita baru",
                "operationId": "ff198a6780504cfd84da3732e9e9a925",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori_id",
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Judul Berita"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Isi berita..."
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "flag": {
                                        "type": "string",
                                        "example": "kegiatan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berita berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Mendapatkan detail berita",
                "operationId": "d0c35e3b939d7260eed68095dcc13b96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID berita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Mengupdate berita",
                "operationId": "eee8c13762ac601365240ab711ed49c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori_id",
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "flag": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berita berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Menghapus berita",
                "operationId": "1c7e28aca0e9810dfb87df5696e6883f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berita berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/news/kategori/{kategori_id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Mendapatkan berita berdasarkan kategori",
                "operationId": "5dc5a638dca8083a07f02ddcf1878611",
                "parameters": [
                    {
                        "name": "kategori_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/news/flag/{flag}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Mendapatkan berita berdasarkan flag",
                "operationId": "f70ae4241b67c835227103e1dd83abf4",
                "parameters": [
                    {
                        "name": "flag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}